Move docker interaction code into own file
